--- jl-tech-test/src/App.jsx
@@ -1,16 +1,0 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import Navbar from './components/Navbar'
-
-function App() {
-  return (
-    <BrowserRouter>
-      <Navbar />
-      <Routes>
-        <Route path='/' element={<h1>Home</h1>} />
-        <Route path='/dashboard' element={<h1>Dashboard</h1>} />
-      </Routes>
-    </BrowserRouter>
-  )
-}
-
-export default App
