+++ jl-test/src/App.js
@@ -1,0 +1,11 @@
+import Navbar from './components/Navbar'
+
+function App() {
+  return (
+    <>
+      <Navbar />
+    </>
+  )
+}
+
+export default App
