+++ jl-test/src/components/Navbar.jsx
@@ -1,0 +1,40 @@
+import React from 'react'
+
+const Navbar = () => {
+  return (
+    <nav className='navbar navbar-expand-lg navbar-dark bg-dark'>
+      <div className='container'>
+        <a className='navbar-brand' href='/'>
+          JL Test
+        </a>
+        <button
+          className='navbar-toggler'
+          type='button'
+          data-bs-toggle='collapse'
+          data-bs-target='#navbarSupportedContent'
+          aria-controls='navbarSupportedContent'
+          aria-expanded='false'
+          aria-label='Toggle navigation'
+        >
+          <span className='navbar-toggler-icon'></span>
+        </button>
+        <div className='collapse navbar-collapse' id='navbarSupportedContent'>
+          <ul className='navbar-nav ms-auto mb-2 mb-lg-0'>
+            <li className='nav-item'>
+              <a className='nav-link active' aria-current='page' href='/'>
+                Home
+              </a>
+            </li>
+            <li className='nav-item'>
+              <a className='nav-link' href='/'>
+                Log In
+              </a>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </nav>
+  )
+}
+
+export default Navbar
