+++ jl-test/src/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import App from './App'
+import 'bootstrap/dist/css/bootstrap.min.css'
+import './index.css'
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+)
